client: allow specifying custom set of headers
Some servers in the wild, e.g. HTTP servers which are proxying a
real `text/event-stream` server API, are not set up to properly
negotiate content types. In such a situation, specifying `Accept: text/event-stream`
may for example cause the request to fail when it would otherwise respond
with a completely valid event stream.

Clearly that is a bug in the server, but it is still valuable to have go-sse
be able to connect to such servers. This change enables that by only
setting default request headers if the HTTP request has none.

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>